refactor: group reader backends
