+++ src/reader/backend/native_async/metered_object_store.rs
